refactor: set up built-in multi validation system
