+++ packages/next-safe-action/src/validation-adapters/index.ts
@@ -1,0 +1,3 @@
+export { zodAdapter } from "./libs/zod";
+
+export * from "./types";
